Extract merge tests from test_fakes
